Reorganize core tests 2
